+++ packages/root/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+- `options.pathname` is a verification function; if it fails, `generic_location` throws.
+
 ## 2.0.0-beta.5
 
 - Rename `locationUtils` to `location_utils`; its API is also now snake cased.