+++ packages/root/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 2.0.0-beta.8
 
 - The `stringify` function from `location_utils` can receive a location object or a string.
 - Remove `AnyLocation` type, add `Hrefable` type.